requirements.txt
asgiref==3.10.0
asttokens==3.0.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.1
joblib==1.5.3
jupyter_core==5.9.1
matplotlib-inline==0.2.1
mysql==0.0.3
mysql-connector-python==9.5.0
mysqlclient==2.2.7
nest-asyncio==1.6.0
numpy==2.4.2
packaging==26.0
pandas==3.0.1
parso==0.8.5
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzmq==27.1.0
qrcode==8.2
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.4
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.6.0
wheel==0.46.3
